# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Grooming CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pets:
    get:
      operationId: listPets
      tags: [pets]
      summary: List pets
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: tenantId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of pets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pet"
    post:
      operationId: createPet
      tags: [pets]
      summary: Create a pet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetInput"
      responses:
        "201":
          description: Created pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
  /pets/{id}:
    get:
      operationId: getPet
      tags: [pets]
      summary: Get a pet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
    patch:
      operationId: updatePet
      tags: [pets]
      summary: Update a pet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetUpdate"
      responses:
        "200":
          description: Updated pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
    delete:
      operationId: deletePet
      tags: [pets]
      summary: Delete a pet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Services
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List service packages
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a service package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Created package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{id}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get a service package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    patch:
      operationId: updatePackage
      tags: [packages]
      summary: Update a service package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageUpdate"
      responses:
        "200":
          description: Updated package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete a service package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /packages/{id}/sell:
    post:
      operationId: sellPackage
      tags: [packages]
      summary: Sell a package (generates appointments + financial entry)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellPackageInput"
      responses:
        "201":
          description: Created appointments and financial entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellPackageResult"

  # Appointments
...
components:
...
  schemas:
...
    Client:
      type: object
      required: [id, tenantId, name, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Pet:
      type: object
      required: [id, clientId, name, size, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        name:
          type: string
        breed:
          type: ["string", "null"]
        size:
          type: string
          enum:
            - mini_curto
            - mini_longo
            - pequeno_curto
            - pequeno_longo
            - medio_curto
            - medio_longo
            - grande_curto
            - grande_longo
            - gigante
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PetInput:
      type: object
      required: [clientId, name, size]
      properties:
        clientId:
          type: integer
        name:
          type: string
        breed:
          type: string
        size:
          type: string
          enum:
            - mini_curto
            - mini_longo
            - pequeno_curto
            - pequeno_longo
            - medio_curto
            - medio_longo
            - grande_curto
            - grande_longo
            - gigante
        notes:
          type: string
    PetUpdate:
      type: object
      properties:
        name:
          type: string
        breed:
          type: string
        size:
          type: string
          enum:
            - mini_curto
            - mini_longo
            - pequeno_curto
            - pequeno_longo
            - medio_curto
            - medio_longo
            - grande_curto
            - grande_longo
            - gigante
        notes:
          type: string

    # Service
...
    ServiceItem:
      type: object
      required: [serviceName, quantity]
      properties:
        serviceName:
          type: string
        quantity:
          type: integer
          minimum: 1
    PriceBySize:
      type: object
      required: [size, price]
      properties:
        size:
          type: string
        price:
          type: number
    Package:
      type: object
      required: [id, tenantId, name, serviceItems, priceBySizes, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        serviceItems:
          type: array
          items:
            $ref: "#/components/schemas/ServiceItem"
        priceBySizes:
          type: array
          items:
            $ref: "#/components/schemas/PriceBySize"
        createdAt:
          type: string
          format: date-time
    PackageInput:
      type: object
      required: [tenantId, name]
      properties:
        tenantId:
          type: integer
        name:
          type: string
        description:
          type: string
        serviceItems:
          type: array
          items:
            $ref: "#/components/schemas/ServiceItem"
        priceBySizes:
          type: array
          items:
            $ref: "#/components/schemas/PriceBySize"
    PackageUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        serviceItems:
          type: array
          items:
            $ref: "#/components/schemas/ServiceItem"
        priceBySizes:
          type: array
          items:
            $ref: "#/components/schemas/PriceBySize"
    SellPackageInput:
      type: object
      required: [tenantId, clientId, petId, startDate, startTime]
      properties:
        tenantId:
          type: integer
        clientId:
          type: integer
        petId:
          type: integer
        startDate:
          type: string
          format: date
        startTime:
          type: string
          description: "HH:MM"
        notes:
          type: ["string", "null"]
    SellPackageResult:
      type: object
      required: [appointments, financialEntry]
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentFull"
        financialEntry:
          $ref: "#/components/schemas/FinancialEntry"

    # Appointment
    AppointmentFull:
      type: object
      required: [id, tenantId, petId, clientId, scheduledDate, status, totalPrice, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        petId:
          type: integer
        clientId:
          type: integer
        serviceId:
          type: ["integer", "null"]
        packageId:
          type: ["integer", "null"]
        scheduledDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [aguardando, em_atendimento, concluido, cancelado]
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
        recurringWeeks:
          type: ["integer", "null"]
        recurringGroupId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        pet:
          $ref: "#/components/schemas/Pet"
        client:
          $ref: "#/components/schemas/Client"
        service:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            name:
              type: string
            price:
              type: number
        package:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            name:
              type: string
            price:
              type: number
...
    FinancialEntry:
      type: object
      required: [id, tenantId, type, description, amount, date, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        type:
          type: string
          enum: [receita, despesa, despesa_fixa]
        description:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        category:
          type: ["string", "null"]
        appointmentId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...